260909_내가 쓴 크루 조회 기능 수정, 알림창 추가
1. 내가 쓴 크루 조회 에러 > 수정
2. 등록, 신청, 취소, 삭제 시 알림창 추가
3. 비밀 댓글의 답댓글 비밀글 처리

+++ src/notice/component/NoticeDetailComponent.jsx
@@ -59,6 +59,9 @@ function NoticeDetailComponent(){
     // 삭제하기 버튼 클릭 시 실행할 이벤트 핸들러 함수
     const deleteNotice = async() => {
 
+        const isConfirm = window.confirm("공지사항을 삭제하시겠습니까?");
+        if (!isConfirm) return;
+
         try{
             const response = await deleteNoticeApi(noticeId);
             if(response.data == "success"){
